+++ app/javascript/react/components/ExportDataModal/EmailInput.tsx
@@ -4,23 +4,19 @@ import { useTranslation } from "react-i18next";
 
 import { TextInput } from "../Modal/Modal.style";
 
-interface ConfirmationMessageProps {
-  message: string;
-}
-
 interface EmailInputProps {
   focusInputRef: React.RefObject<HTMLInputElement>;
   value: string;
   onChange: (event: React.ChangeEvent<HTMLInputElement>) => void;
 }
 
-const { t } = useTranslation();
-
 const EmailInput: React.FC<EmailInputProps> = ({
   focusInputRef,
   value,
   onChange,
 }) => {
+  const { t } = useTranslation(); // Move useTranslation inside the component
+
   return (
     <TextInput
       type="email"
@@ -33,19 +29,8 @@ const EmailInput: React.FC<EmailInputProps> = ({
   );
 };
 
-const MessageContainer = styled.div`
-  margin-top: 1rem;
-  text-align: center;
+const RedErrorMessage = styled.span`
+  color: red;
 `;
 
-const ConfirmationMessage: React.FC<ConfirmationMessageProps> = ({
-  message,
-}) => {
-  return (
-    <MessageContainer>
-      <h2>{message}</h2>
-    </MessageContainer>
-  );
-};
-
-export { EmailInput, ConfirmationMessage };
+export { EmailInput, RedErrorMessage };
